Solve issue #1
Doxygen documentation added.
Doc configuration files added.
Structure accordingly to arduino specification.
