--- README.md
@@ -1,14 +1,0 @@
-Nanoshield_RTC
-==============
-
-This is the Arduino software library to access the RTC functionality of the RTCMem Nanoshield (PCF8563).
-
-This is used to read and write the date and time to the RTCMem Nanoshield.
-
-To install, just copy it under &lt;arduinosketchfolder&gt;/libraries/
-
----
-Copyright (c) 2013 Circuitar  
-All rights reserved.
-
-This software is released under an MIT license. See the attached LICENSE file for details.
