+++ README.rst
@@ -1,0 +1,35 @@
+Nanoshield_RTC
+==============
+
+This is the Arduino software library to access the RTC functionality of the RTCMem Nanoshield.
+
+* Source code: https://github.com/circuitar/Nanoshield_RTC
+* Documentation: http://nanoshield-rtc.readthedocs.org/
+* Reference board: `Nanoshield RTCMen`_ from Circuitar_
+
+Library features include:
+
+* Read date and time from the RTCMem Nanoshield
+* Write date and time to the RTCMem Nanoshield
+
+To install, just click **Download ZIP** and install it using **Sketch > Include Library... > Add .ZIP Library** in the Arduino IDE.
+
+----
+
+The following examples_ are provided:
+
+- SimpleClock_ serial port clock application using the RTC Nanoshield.
+- SimpleClockLCD_ clock application using the RTC Nanoshield and the LCD Nanoshield.
+
+.. _`Nanoshield RTCMem`: https://www.circuitar.com.br/nanoshields/modulos/rtcmem/
+.. _Circuitar: https://www.circuitar.com.br/
+.. _examples: https://github.com/circuitar/Nanoshield_RTC/tree/master/examples
+.. _SimpleClock: https://github.com/circuitar/Nanoshield_RTC/blob/master/examples/SimpleClock/SimpleClock.ino
+.. _SimpleClockLCD: https://github.com/circuitar/Nanoshield_RTC/blob/master/examples/SimpleClockLCD/SimpleClockLCD.ino
+
+----
+
+Copyright (c) 2013 Circuitar  
+All rights reserved.
+
+This software is released under an MIT license. See the attached LICENSE file for details.
