+++ extras/docs/index.rst
@@ -1,0 +1,26 @@
+.. toctree::
+   :maxdepth: 2
+
+.. include:: ../../README.rst
+   :end-before: ----
+
+Class Documentation
+===================
+
+.. doxygenclass:: DS1307
+   :project: Nanoshield_RTC
+   :members:
+
+.. doxygenclass:: Nanoshield_RTCPlus
+   :project: Nanoshield_RTC
+   :members:
+
+.. doxygenclass:: Nanoshield_RTC
+   :project: Nanoshield_RTC
+   :members:
+
+----
+
+This documentation was built using ArduinoDocs_.
+
+.. _ArduinoDocs: http://arduinodocs.readthedocs.org
