+++ extras/docs/requirements.txt
@@ -1,0 +1,1 @@
+breathe
