+++ library.properties
@@ -1,0 +1,9 @@
+name=Nanoshield_RTC
+version=1.0
+author=Luis Chavier <lfchavier@circuitar.com.br>
+maintainer=Luis Chavier <lfchavier@circuitar.com.br>
+sentence=This is the Arduino software library to access the RTC functionality of the RTCMem Nanoshield.
+paragraph=Provide read and write date and time to RTCMem Nanoshield.
+category=Timing
+url=https://github.com/circuitar/Nanoshield_RTC
+architectures=*
